src/main/java/br/com/luishenrique/bonus_funcionario/service/BonusService.java
package br.com.luishenrique.bonus_funcionario.service;

import br.com.luishenrique.bonus_funcionario.model.Funcionario;
import java.math.BigDecimal;
import java.math.RoundingMode;

public class BonusService {

    public BigDecimal calcularBonus(Funcionario funcionario) {
        BigDecimal valor = funcionario.getSalario().multiply(new BigDecimal("0.1"));
        if (valor.compareTo(new BigDecimal("1000")) > 0) {
            valor = BigDecimal.ZERO;
        }
        return valor.setScale(2, RoundingMode.HALF_UP);
    }
}
